feat: cambio de sheet a pagina para realizar el update
